Comments on egress primitives and implementations. Minor refactors for organization and readability.
