+++ src/Microsoft.Diagnostics.Monitoring/Egress/EgressProviderOptions.cs
@@ -4,8 +4,15 @@
 
 namespace Microsoft.Diagnostics.Monitoring.Egress
 {
+    /// <summary>
+    /// Egress provider options common to all egress providers.
+    /// </summary>
     internal abstract class EgressProviderOptions
     {
+        /// <summary>
+        /// Buffer size used when copying data from an egress callback returning a stream
+        /// to the egress callback that is provided a stream to which data is written.
+        /// </summary>
         public int? CopyBufferSize { get; set; }
     }
 }
