+++ src/Microsoft.Diagnostics.Monitoring/Egress/FileSystem/FileSystemEgressProviderOptions.cs
@@ -6,12 +6,22 @@ using System.ComponentModel.DataAnnotations;
 
 namespace Microsoft.Diagnostics.Monitoring.Egress.FileSystem
 {
+    /// <summary>
+    /// Egress provider options for file system egress.
+    /// </summary>
     internal class FileSystemEgressProviderOptions :
         EgressProviderOptions
     {
+        /// <summary>
+        /// The directory path to which the stream data will be egressed.
+        /// </summary>
         [Required]
         public string DirectoryPath { get; set; }
 
+        /// <summary>
+        /// The directory path to which the stream data will initially be written, if specified; the file will then
+        /// be moved/renamed to the directory specified in <see cref="FileSystemEgressProviderOptions.DirectoryPath"/>.
+        /// </summary>
         public string IntermediateDirectoryPath { get; set; }
     }
 }
