+++ src/Tools/dotnet-monitor/Egress/Configuration/EgressConfigurationExtensions.cs
@@ -1,0 +1,19 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Microsoft.Extensions.Configuration;
+
+namespace Microsoft.Diagnostics.Tools.Monitor
+{
+    internal static class EgressConfigurationExtensions
+    {
+        /// <summary>
+        /// Get the Egress configuration section from the specified configuration.
+        /// </summary>
+        public static IConfigurationSection GetEgressSection(this IConfiguration configuration)
+        {
+            return configuration.GetSection(EgressOptions.ConfigurationKey);
+        }
+    }
+}
