+++ src/Tools/dotnet-monitor/Egress/Configuration/EgressOptions.cs
@@ -7,13 +7,23 @@ using System.Collections.Generic;
 
 namespace Microsoft.Diagnostics.Tools.Monitor
 {
+    /// <summary>
+    /// Configuration options for specifying egress providers.
+    /// </summary>
     internal class EgressOptions
     {
         public const string ConfigurationKey = "Egress";
 
+        /// <summary>
+        /// Mapping of egress provider names to egress provider implementations.
+        /// </summary>
         public Dictionary<string, ConfiguredEgressProvider> Providers { get; }
             = new Dictionary<string, ConfiguredEgressProvider>(StringComparer.OrdinalIgnoreCase);
 
+        /// <summary>
+        /// Mapping of keyed values, typically used for naming a secret such as a storage account
+        /// key or shared access signature rather than embedding values directly in the egress provider options.
+        /// </summary>
         public Dictionary<string, string> Properties { get; }
             = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
     }
